refactor(star): split StarPage into per-state modules
StarPage was ~900 lines. Extract each concern into components/subscription/star/:
console shell + buttons (StarConsole), CenterReadout, and one file per flow
state (Overview/Method/Pay/Success/Manage/RedeemPane), plus pure helpers
(meta: PERKS/monthsKey/primaryEntitlement/Step; redeem-code: code parsing).
StarPage is now ~200 lines of pure orchestration. No behaviour change.
