+++ desktop/src/components/subscription/star/ManagePane.tsx
@@ -1,0 +1,95 @@
+import { useMutation, useQuery } from '@tanstack/react-query';
+import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { toast } from 'sonner';
+import { payApi } from '../../../lib/pay-client';
+import { requestPremiumRecheck } from '../../../lib/premium-cache';
+import { queryClient } from '../../../lib/query-client';
+import { passDate } from '../../../lib/star-format';
+import { DiscordCard } from '../DiscordCard';
+import { primaryEntitlement } from './meta';
+import { GhostBtn } from './StarConsole';
+
+/** Manage — Discord link, auto-renew (cancel-only), source; extend / redeem. */
+export const ManagePane = memo(function ManagePane({
+  onExtend,
+  onRedeem,
+}: {
+  onExtend: () => void;
+  onRedeem: () => void;
+}) {
+  const { t } = useTranslation();
+  const SUB_KEY = ['pay', 'subscription'] as const;
+  const { data, isLoading } = useQuery({
+    queryKey: SUB_KEY,
+    queryFn: payApi.subscription,
+    staleTime: 30_000,
+  });
+  const ent = data ? primaryEntitlement(data.entitlements) : null;
+  const source = ent?.source ?? '';
+  const endsAt = ent?.ends_at ?? data?.premium_until ?? 0;
+  const autoRenew = !!ent?.auto_renew && !ent?.canceled;
+
+  const cancel = useMutation({
+    mutationFn: () => payApi.cancel(source),
+    onSuccess: () => {
+      toast.success(t('starpass.manage.canceled'));
+      void queryClient.invalidateQueries({ queryKey: SUB_KEY });
+      requestPremiumRecheck();
+    },
+    onError: () => toast.error(t('starpass.manage.cancelError')),
+  });
+
+  return (
+    <div className="flex flex-col gap-2.5">
+      <DiscordCard discord={data?.discord ?? null} />
+
+      <div className="flex items-center justify-between gap-3 rounded-[14px] border border-white/[0.08] bg-white/[0.03] px-4 py-3">
+        <div>
+          <div className="text-[13px] font-medium text-white/85">
+            {t('starpass.manage.autoRenew')}
+          </div>
+          <div className="mt-0.5 text-[11.5px] text-white/40">
+            {autoRenew
+              ? t('starpass.manage.autoRenewOn', { date: passDate(endsAt) })
+              : t('starpass.manage.autoRenewOff')}
+          </div>
+        </div>
+        <button
+          type="button"
+          role="switch"
+          aria-checked={autoRenew}
+          disabled={!autoRenew || cancel.isPending || isLoading || !source}
+          onClick={() => autoRenew && !cancel.isPending && cancel.mutate()}
+          className="relative h-6 w-11 shrink-0 cursor-pointer rounded-full border border-white/[0.12] transition-all duration-300 disabled:cursor-default"
+          style={{
+            background: autoRenew
+              ? 'color-mix(in srgb, var(--color-accent) 22%, transparent)'
+              : 'rgba(255,255,255,0.04)',
+          }}
+        >
+          <span
+            className="absolute top-0.5 size-5 rounded-full transition-all duration-300"
+            style={{
+              left: autoRenew ? 22 : 2,
+              background: autoRenew ? 'var(--color-accent)' : 'rgba(255,255,255,0.4)',
+            }}
+          />
+        </button>
+      </div>
+
+      <div className="flex items-center justify-between gap-3 rounded-[14px] border border-white/[0.08] bg-white/[0.03] px-4 py-3">
+        <div>
+          <div className="text-[13px] font-medium text-white/85">{t('starpass.manage.source')}</div>
+          <div className="mt-0.5 text-[11.5px] text-white/40">{t('starpass.manage.sourceSub')}</div>
+        </div>
+        <span className="font-mono text-[13px] text-white">{source || '—'}</span>
+      </div>
+
+      <div className="mt-1 flex flex-wrap gap-3">
+        <GhostBtn onClick={onExtend}>{t('starpass.manage.extend')}</GhostBtn>
+        <GhostBtn onClick={onRedeem}>{t('starpass.manage.redeem')}</GhostBtn>
+      </div>
+    </div>
+  );
+});
